Readying for first release
There's still a lot of things that should be cleaned up,
a lot of code is actually duplicated because "let's rush
for a first release".

Still everything is working fine, the error messages are
pretty clear, however if the thing fails, (because you
setup a folder that's then deleted) it still let a lot of
garbages inside the hierarchy...

So, yeah, it works, but could require some additional
work.

Still, while doing this project, it became more and more
clear that it's possible to setup some kind of interface
where you select the "skills" you want to add to your
character and then add pre-packaged animations/menu items
in a more meaningful way than this.

Signed-off-by: Voyage <voyage@miouyouyou.fr>