Major code cleanup
The main UI will be the next one.

The system now handles the ability to lock down items from
world origin, which allows you to either locate the world
origin everytime, or make world overlays with your avatar.

Helpers have been moved into their own library. This doesn't
change anything from a user overview, however this might
introduce some changes to the packaging in a near future.

There's multiple reasons for that.

One is that I'd like to reuse these helpers in others projects,
since they help me develop way faster, when it comes to
generating animations and handling unity hierarchy.

The other is that with VRChat move to packaging systems,
I might be able to actually pull the helpers as a package
dependency.

I'd also like to convert this to a package useable from the
VRChat Creator Companion at some point.

Still, a good part of the cleanup is done.
Next phase is the interfaces code, since the whole polymorphism
didn't help much here in the end.
And also managing the translations, in order to have a UI
in Japanese for japanese people at least.

Signed-off-by: Voyage <voyage@miouyouyou.fr>