Fixed: Error message for menus with too much controls
Both the Japanese and English versions were unreadable, IMHO.

Also, turns out that I was dead wrong on the "Reset avatar" button.
It's added anyway, even if you fill the menu with 8 controls.

So the maximum amount of controls allowed in the main menu are now :
* 7 (8-1) controls for Constraints
* 6 (8-2) for Particles.
I could make the limit 7 for Particles too by moving the two added
buttons to a submenu, and adding this submenu.
I'll see if people are really hindered by this limit, since this
requires additional tests, and remaking parts of the presentation
videos.

Also the error message now comes with a new button, allowing you to
quickly open and edit the Expression Menu in question.

Signed-off-by: Voyage <voyage@miouyouyou.fr>