Dealing with parent constraints setup on the object
Some people want to setup objects that are Constraines to
the avatar being equipped.
This lead to the item following the avatar, even when enabling
"World Lock".

So the tool now automatically disables constraints components setup
on the objects during "World Lock".

Disabling constraints is still a bit rough, though, since it
disable all constraints, even those that have zero link to the
avatar.

I'll add more fine handling later.

Also, if the constraints sources of the object to equip link to
a Transform of the main avatar, these links are fixed in order to
point to the same Transform on the avatar copy.

Meaning that if you link your sword to the main avatar hands, they
will be linked to the duplicated avatar hands once equipped using
the Constraint tools.

Signed-off-by: Voyage <voyage@miouyouyou.fr>