Reset to default settings when pressing Reset
And removed the videos from the folder, at the moment.
When encoded to WebM, Unity borks when trying to import them.

These video still weight a lot compared to the main
code, and can be viewed online anyway.

I'm not a fan of removing offline help, but VRChat Avatar
development is inherently linked to a working Internet connection,
and these video are not essential, so they can be stored
online.

Signed-off-by: Voyage <voyage@miouyouyou.fr>