Change the "Constraints disabling" option name
To remove any ambiguity. Now people should clearly
understand that constraints are only disabled when
locking the item, when the option is activated.

Previously it sounded like constraints would get
disabled at any time.

Also made Constraint Sources remapping more flexible,
so that you can equip an avatar copy with an item
constrained to the main avatar.
When doing this, the constraint source is remapped
correclty to the same transform inside the avatar
copy hierarchy.

Signed-off-by: Voyage <voyage@miouyouyou.fr>