Avoid scaling down the item
Scaling down the container is the way to go.
The reason being that scaling down the item makes it
impossible to get the actual scale when trying to setup
the Stations Proxies objects.

The reason why most of these proxies worked before
was that while the main object was scaled down to 0,
the children were not and the stations were often put
on the children.

This takes care of the last bug I've been able to
detect. So long, so far, everything works now.

So readying up for release !

Signed-off-by: Voyage <voyage@miouyouyou.fr>