Reverting the "optimization"
Because Unity ignores any animation track played while
the user is culled, and because I need at least one frame
between :
* The time the container is enabled with its Parent Constraint
  Active
* The moment that Parent Constraint is disabled

I cannot "optimize" the avatar by just disabling the container.

Enable GameObject and Disable Parent Constraint have to be
executed on different frames.
So in the Animation Controller Layer, the setup will be :

[OFF]
  ↓ (Lock0 : true)
[Enable GameObject]
  ↓ (Lock0 : true / No Condition / Whatever)
[Enable Parent Constraint]

If you're not looking at the player while Lock0 gets to true,
the player will get culled.
During the culling, the animation Enable GameObject will play
but its content will be ignored.
Same thing for Enable Parent Constraint.

When you look back at the player, the culling will stop and
Unity will resume the animations from the last played animation
frame, which is the last frame of "Enable Parent Constraint".
(These animations don't loop by the way, since there's no reason
to do so).

So, in result : You won't see the object !

Think about it !
The entire Enable GameObject is ignored. It's played but the content
is completely ignored.

Now you could create ONE animation with Enable GameObject and
Enable Parent Constraint at the same time.
And you can make it so that the Parent Constraint is enabled one
frame and disabled on the next.

But that doesn't solve the problem, since Unity resumes the LAST FRAME.
Not the entire animation. Just the last frame. (Unless you're loopping
or inside a very long animation, which is not the case here).

So, at best, your item will be displayed but locked at 0,0,0 or at the
last place where you saw the player.

Now, keeping the 3 constraints (Rot/Pos/Parent) always active doesn't
solve all the problems, but at least resolve the fact that people
will AT LEAST  see the object near where you are if they were not
looking at your during the activation.

Signed-off-by: Voyage <voyage@miouyouyou.fr>