Quick and dirty fix for issue #1
Until I find a more robust way to deal with this.
I didn't expect VRChat to ACTUALLY CHANGE the array
depending on the avatar type.
I thought, this SDK would just HIDE the useless layers,
not redo the entire array and mess up the indices !

Anyway, the thing here is that :
* I'm reimplementing "Find" in an array, which I hate.
  There might be a smarter way to do this in C#. I'll have
  to look it up.
* What happens if VRChat suddenly decides that
  "Avatar of type Y has no FX layer !".
  I'll need some code for potentiel corner cases...

Also, the fix is just tested in Unity but not in VRChat yet

Signed-off-by: Voyage <voyage@miouyouyou.fr>