Fixed anomaly #1
This time it's fixed in a slightly better way...
Though I can't really test it fully.

Still, I tested with Generic avatar and normal avatars
and it's working fine.
One person tested the last fix. I'll package this version
and see if the next version is working fine too.

If that's the case, I'll close the isssue and release
the fix on Booth too.

Signed-off-by: Voyage <voyage@miouyouyou.fr>