Migrating towards VRC Constraints
I still have to clean the code and make it way simpler
that it is currently.

The new VRChat Constraints are really well made, user-wise.
SDK wise, there's some "WTF IS THIS" parts.

Still, the "World Lock" check just works. There's some nice
tricks with it too that I haven't took the time to play with.
Setting the transform, you can do some nifty tricks.

I still think I'll add a toggle system.
Sure, it's an additional layer that might not be needed, but
you generally want to toggle the item, whether it's a sword,
a table, or anything you would worldlock.

Signed-off-by: Voyage <voyage@miouyouyou.fr>