Toggles with your VRC Constraints !
When you have a sword, or a weapon, or some furniture,
you don't want to "just world lock", you want to toggle
the item too.

So here it is !

World lock and toggles !

And if you want the previous behaviour, it's still available.

I had to test so many combinations due to this, but
here we are !

I still need to test the old mechanisms before release.

Signed-off-by: Voyage <voyage@miouyouyou.fr>