Fixed various bugs due to name conflicts
Meaning that you can setup the same item 10 times
in a row, or you can setup an item with the same
name as one actually setup on the avatar.

The menu was also dealt with and everything was tested
rather extensively.
So I'm pretty confident now about releasing the
software.

The UI is still garbage, but that will be for the
next upgrade.

Signed-off-by: Voyage <voyage@miouyouyou.fr>