[VRCConstraints] Fixed the recent bugs
So, I got a report about a really weird anomaly
the previous week, where people could lock (freeze)
the item once, unlock it, but locking it again
would make it go to the original lock location.

It turns out that if you don't define a Source
for the VRC Parent Constraint used to "Freeze"
(World Lock) the item, the item will just go back
to its first frozen location after defining it.

It's a NEW bug but I don't think VRChat will fix it
anyway.

Fixing this bug actually led me to use all the
options of the VRC Constraint, which decreased the
complexity of some parts of the code.

I also fixed two bugs :

1. Fixed an issue where an item constrained to an
avatar location, would not be adapted to the copy
due to the object being inactive at first.
That was due to GetComponentsInChildren not looking
into inactive objects by default...

2. Fixed an issue where setting up a single button
to output and lock a default-hidden object with a
VRC Parent Constraint on it would not show up.
This issue is REALLY weird and my only solution was
to actually remove the VRC Constraints of the object
when using such a combination of options.
So basically :
* [ ] Show the item by default
* [ ] Separate the buttons
* [x] Disable the constraints on lock
Will lead to the deletion of the Constraints if
present.
In such a configuration, these constraints make zero
sense anyway.
When not locking the object, the object is inactive.
And when it's active due to pressing the main button,
the constraints are disabled.

Signed-off-by: Voyage <voyage@miouyouyou.fr>